ricode/requirements.txt
pyspark==3.4.0
delta-spark==2.4.0
selenium==4.10.0
matplotlib==3.8.0
seaborn==0.12.2
beautifulsoup4==4.12.2
numpy==1.24.4
pandas==2.1.0